Add namespaced permissions with cluster read
Makes sense if we are going to allow customization of
aggregator permissions that we provide a few useful options.

In the originating ticket, the users wanted a strictly namespaced
permissions set.

It also makes sense, to offer a namespaced permission set that comes
with cluster reads. This allows us to run daemonsets (requires listing
nodes) and allows plugins to check things like kube-system pods.

In this permission mode, e2e tests will start (checks nodes are up
and kube-system pods are ready) but will fail since tests require that
namespaces get created (which it lacks permissions for).

Signed-off-by: John Schnake <jschnake@vmware.com>