Add conformance-lite mode
The goal of this mode is to maximize coverage while minimizing
test time. I was able to get 257 tests in 334s.

This is just done from looking at old runs and sorting the tests
by time and avoiding running the slowest ones. We do not specify
the fastest ones because there are so many of them.

Due to the length of the skip flag, I added a truncated/verbose option
for `sonobuoy modes` to avoid a bad user experience.

Signed-off-by: John Schnake <jschnake@vmware.com>