Add e2e command for dry-run help
This command simply takes the e2e version and a focus/skip value
and will get the list of e2e tests which would be run. It can then
print the tests, their tags, or their tags and the number of times
they appeared.

It can take data from stdin, our github location, or the hardcoded
values it was built with (same as github at the time of release).

Signed-off-by: John Schnake <jschnake@vmware.com>