Adjust progress messages to allow for appending data
In the parallel ginkgo case, each thread doesn't know the
totals of how many tests will be run or have been run.

As a result we need to be able to handle partial updates such
as just saying one more test passed/failed.

This adds a few fields to the data type and helper functions/tests
to ensure that the appending works as intended.

Upstream change to utilze this will be next.

Signed-off-by: John Schnake <jschnake@vmware.com>