Worker sidecars get plugin env vars
To allow users a way to specify env vars for the workers,
and therefore customize behavior, we copy the main set of
env vars from the plugin to the worker container. In the case
of conflicts, we prioritize our preset value to avoid breaking
the worker (e.g. if they try and override the address of the
aggregator).

Signed-off-by: John Schnake <jschnake@vmware.com>