Modify reader code to work on unzipped results as well
The code for the results reader was made to read the tarball without
opening it but it should also work on unzipped results. This allows
more reuse of the methods, a centralized place to put results structure
information, and will allow us to expand `results` to process an unzipped
directory as well.

Since the extra work was minimal and it served as a good test, I also
modified the results command to accept a directory.

Signed-off-by: John Schnake <jschnake@vmware.com>