Fix bug where 'retrieve' fails if filepath diectories DNE
The logic takes a path and a file; if the file has path values in it
that do not exist, then os.Create was failing but the MkdirAll logic
wasn't looking for paths in that.

To fix this, I just reconcile the directory/filename to get the full path
and process it accordingly sooner.

Fixes #1689

Signed-off-by: John Schnake <jschnake@vmware.com>