Moved the logic for gathering logs to before plugin cleanup
The cleanup logic must occur after queries or else pod logs
won't be gathered (it will be a race, at best).

There is a test for this code but I think it was a race that
worked in our favor so it would pass in CI but with real plugins
it would not. Will follow up with other test modifications but want
to get this fix in asap.

Signed-off-by: John Schnake <jschnake@vmware.com>