Only apply imagePullPolicy to plugins if instructed (#1705)
Adds a new option which allows the user to specify whether or
not to apply the imagePullPolicy to all plugins and override
their stated values.

This is important since even the default value was overwriting
explicit values in the plugins leading to confusing behavior.

Techinically a breaking change for someone who was running
sonobuoy CLI with the --image-pull-policy flag to overwrite explicit
values. They will now need to add the --force-image-pull-policy flag
to obtain the same behavior. This makes all modifications to plugins
explicit via some option rather than this one being implicit.

Fixes #1652

Signed-off-by: John Schnake <jschnake@vmware.com>