Add CLI docs to the site
Uses cobra to generate markdown docs for the command tree.

Known issue to resolve is that the default image used is based on
the sonobuoy version running the command. Right now that is just
a locally build version which is marked with a sha and as dirty.

Need a little think about the best way to resolve that since the
docs wont be changed on every commit, and the 'main' branch of sonobuoy
would be targeting different images. We just need to make sure that
tagged releases have the right version when those docs are published.

Fixes #1727

Signed-off-by: John Schnake <jschnake@vmware.com>