Fix skopeo login for image push issue
Locally I reproduced the bug and was able to fix it by changing the
skopeo login to the $REGISTRY and not to the $DOCKERHUB_USER.

This begs the question: why did it work in the past? Unsure of the root
cause but this seemed to remedy it.

Also added debug info so that we can better diagnose the issue in the
future.

Fixes #1730

Signed-off-by: John Schnake <jschnake@vmware.com>