Move to registry.k8s.io
The community is trying to reduce traffic to k8s.gcr.io
and is asking that people move to registry.k8s.io whenever
possible. Currently there is a proxy that will redirect you
so it won't break if we dont updata right away, but since
the change is so close, it is advised that we move now.

Fixes #1734

Signed-off-by: John Schnake <jschnake@vmware.com>