POC for local plugins via crashdiagnostics
This is a POC branch to run local plugins. This gives us a way to
integrate the functionality of crash-diagnostics into sonobuoy.

TODOs:
 - ensure if running against broken API, the queries fail fast rather than
timing out after X minutes
 - consider whether or not scripts/args should be embedded in the plugin yaml
itself. I think either is OK, but personally embedding them in the plugin yaml
makes it harder to edit because of the extra indenting. The script/args are already
persisted in the result tarball once run. However, not embedding it in the plugin yaml
makes it harder to share a plugin (since you have to share 3 files instead of 1).